Functionnal staatic element factory

+++ model/src/main/java/model/element/Position.java
@@ -1,27 +1,27 @@
 package model.element;
 
 public class Position implements IPosition {
-	private int x;
-	private int y;
+	private float x;
+	private float y;
 	
-	public Position(int x,int y) {
+	public Position(float x,float y) {
 		this.setX(x);
 		this.setY(y);
 	}
 	
-	public int getX() {
+	public float getX() {
 		return x;
 	}
 	
-	public void setX(int x) {
+	public void setX(float x) {
 		this.x = x;
 	}
 	
-	public int getY() {
+	public float getY() {
 		return y;
 	}
 	
-	public void setY(int y) {
+	public void setY(float y) {
 		this.y = y;
 	}
 }
